+++ src/sll/include/sll/_internal/env.h
@@ -1,0 +1,10 @@
+#ifndef __SLL__INTERNAL_ENV_H__
+#define __SLL__INTERNAL_ENV_H__ 1
+
+
+
+void _init_env(void);
+
+
+
+#endif